% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leer_actual.R
\name{leer_actual}
\alias{leer_actual}
\title{Lee varias jornadas de la temporada actual}
\usage{
leer_actual(desde, hasta)
}
\arguments{
\item{desde}{Jornada inicial}

\item{hasta}{Jornada final}
}
\description{
Lee varias jornadas de la temporada actual
}
