% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leer_jornada.R
\name{leer_jornada}
\alias{leer_jornada}
\title{Lee una jornada de la temporada actual}
\usage{
leer_jornada(jornada)
}
\arguments{
\item{jornada}{Número de jornada}
}
\value{
Data frame con los 14 partidos
}
\description{
Lee una jornada de la temporada actual
}
