% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leer_pronostico.R
\name{leer_pronostico}
\alias{leer_pronostico}
\title{Lee la jornada que está por jugarse}
\usage{
leer_pronostico(jornada)
}
\arguments{
\item{jornada}{Número de jornada}
}
\description{
Lee la jornada que está por jugarse
}
