% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leer_temporada.R
\name{leer_temporada}
\alias{leer_temporada}
\title{Lee una temporada completa (histórica)}
\usage{
leer_temporada(temporada, max_jornada)
}
\arguments{
\item{temporada}{Año de la temporada (2020 = 2020/21)}

\item{max_jornada}{Número total de jornadas de esa temporada}
}
\description{
Lee una temporada completa (histórica)
}
