% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quinielaR.R
\docType{package}
\name{quinielaR}
\alias{quinielaR-package}
\alias{quinielaR}
\title{quinielaR: Paquete para descargar y procesar datos de La Quiniela}
\description{
Este paquete permite descargar temporadas históricas, jornadas actuales
y pronósticos de La Quiniela, además de limpiar y estructurar los datos.
}
